Update Multiply by 50
'''
///
Description

You are given a number stored in a variable, with the namenumber

Multiply the number stored in the variablenumberby 50, and print the result obtained


Input
The first and the only line of the input contains the number stored in the variablenumber

Constraints

1 <= N <= 20


Output
Print a single number, the result obtained by multiplying it by 50


Sample Input 1 

2
Sample Output 1

100
Hint

In the sample test case, the value stored in the variablenumberis 2. After, multiplying by 50, the answer obtained is 100. Therefore, the answer obtained is 100.
///
'''